requirements.txt
# --- Web / API ---
Flask==3.0.2
gunicorn==21.2.0
requests>=2.31.0
python-dotenv>=1.0.1

# --- IA / NLP ---
openai==1.51.0
nltk==3.9.1
scikit-learn==1.5.1
joblib==1.4.2

# --- PDF / Extração de texto ---
PyPDF2==3.0.1
PyMuPDF==1.24.10   # mais rápido/robusto para PDF que não seja só imagem

# --- Suporte a encoding / charset ---
chardet>=5.2.0

# --- (Opcional) OCR para PDFs escaneados/imagem ---
# pytesseract==0.3.13
# ocrmypdf==16.4.1
# Pillow>=10.3.0
# -> Requer instalar Tesseract no sistema (apt-get install tesseract-ocr)
